{us10} be get component per machine

+++ Data-API-Application/src/main/java/Q3Project/DataAPIApplication/Model/MachineMonitoringPoorten.java
@@ -20,7 +20,7 @@ public class MachineMonitoringPoorten {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id")
-    private int machineMonitoringPoortenId;
+    private long machineMonitoringPoortenId;
 
     @Column(name = "board")
     private int board;
