+++ Data-API-Application/src/main/java/Q3Project/DataAPIApplication/Model/Role.java
@@ -20,7 +20,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "role_id")
-    private int roleId;
+    private long roleId;
 
     @Column(name = "name")
     private String name;
